Setting for character encoding of source files.
Adds a feature that set the character encoding of source files.
Currently, character encodings of source files and the platform
have to be matched.

This change adds two ways to set the character encoding of
source files.
1) "report.source.encoding" property as EMMA report
   generation property.
2) "sourceencoding" attribute of html element for Ant.
If the character encoding of source files is not set,
the default encoding of the platform ("file.encoding"
property) is set. If "file.encoding" property is also
not defined, UTF-8 is set.

Currently, the character encoding of source files and
the platform have to be matched. Because the current HTML report
generator uses "file.encoding" property of JRE while reading
source files.

Change-Id: I43cc5d093cdcc9dc157e62146b9609f282d54e07